feat :: character command query 분리
