+++ src/main/java/windeath44/server/memorial/domain/character/controller/CharacterQueryController.java
@@ -1,7 +1,7 @@
 package windeath44.server.memorial.domain.character.controller;
 
 import windeath44.server.memorial.domain.character.dto.response.CharacterResponse;
-import windeath44.server.memorial.domain.character.service.CharacterService;
+import windeath44.server.memorial.domain.character.service.CharacterQueryService;
 import windeath44.server.memorial.global.dto.CursorPage;
 import windeath44.server.memorial.global.dto.ResponseDto;
 import windeath44.server.memorial.global.util.HttpUtil;
@@ -15,25 +15,25 @@ import java.util.List;
 @RequiredArgsConstructor
 @RequestMapping("/animes/characters")
 public class CharacterQueryController {
-    private final CharacterService characterService;
+    private final CharacterQueryService characterQueryService;
 
     @GetMapping
     public ResponseEntity<ResponseDto<CursorPage<CharacterResponse>>> findAll(@RequestParam(value = "cursorId", required = false) Long cursorId, @RequestParam int size) {
-        CursorPage<CharacterResponse> characterResponses = characterService.findAll(cursorId, size);
+        CursorPage<CharacterResponse> characterResponses = characterQueryService.findAll(cursorId, size);
         ResponseDto<CursorPage<CharacterResponse>> responseDto = HttpUtil.success("find characters", characterResponses);
         return ResponseEntity.ok(responseDto);
     }
 
     @GetMapping("/{character-id}")
     public ResponseEntity<ResponseDto<CharacterResponse>> findById(@PathVariable("character-id") Long characterId) {
-        CharacterResponse characterResponse = characterService.find(characterId);
+        CharacterResponse characterResponse = characterQueryService.find(characterId);
         ResponseDto<CharacterResponse> responseDto = HttpUtil.success("find character", characterResponse);
         return ResponseEntity.ok(responseDto);
     }
 
     @GetMapping("/search/anime")
     public ResponseEntity<ResponseDto<CursorPage<CharacterResponse>>> findIdsByAnimeId(@RequestParam("animeId") List<Long> animeId, @RequestParam(value = "cursorId", required = false) Long cursorId, @RequestParam int size) {
-        CursorPage<CharacterResponse> characterResponse = characterService.findByAnime(animeId, size, cursorId);
+        CursorPage<CharacterResponse> characterResponse = characterQueryService.findByAnime(animeId, size, cursorId);
         ResponseDto<CursorPage<CharacterResponse>> responseDto = HttpUtil.success("find character by anime id", characterResponse);
         return ResponseEntity.ok(responseDto);
     }
@@ -47,28 +47,28 @@ public class CharacterQueryController {
             @RequestParam(value = "size") int size,
             @RequestParam(value = "memorialState", required = false) String memorialState // NOT_MEMORIALIZING, MEMORIALIZING
     ) {
-        CursorPage<CharacterResponse> characterResponses = characterService.findAllIntegrated(name, animeId, deathReason, memorialState, cursorId, size);
+        CursorPage<CharacterResponse> characterResponses = characterQueryService.findAllIntegrated(name, animeId, deathReason, memorialState, cursorId, size);
         ResponseDto<CursorPage<CharacterResponse>> responseDto = HttpUtil.success("find character ids integrated", characterResponses);
         return ResponseEntity.ok(responseDto);
     }
 
     @GetMapping("/search/death-reason")
     public ResponseEntity<ResponseDto<CursorPage<CharacterResponse>>> findIdsByDeathReason(@RequestParam("deathReason") String deathReason, @RequestParam(value = "cursorId", required = false) Long cursorId, @RequestParam int size) {
-        CursorPage<CharacterResponse> characterResponses = characterService.findAllByDeathReason(deathReason, cursorId, size);
+        CursorPage<CharacterResponse> characterResponses = characterQueryService.findAllByDeathReason(deathReason, cursorId, size);
         ResponseDto<CursorPage<CharacterResponse>> responseDto = HttpUtil.success("find character ids by death reason", characterResponses);
         return ResponseEntity.ok(responseDto);
     }
 
     @GetMapping("/search/characterIds")
     public ResponseEntity<ResponseDto<List<CharacterResponse>>> findCharacterResponsesByCharacterIds(@RequestParam List<Long> characterIds) {
-        List<CharacterResponse> characterResponseList = characterService.findByCharacterIds(characterIds);
+        List<CharacterResponse> characterResponseList = characterQueryService.findByCharacterIds(characterIds);
         ResponseDto<List<CharacterResponse>> responseDto = HttpUtil.success("find characters by character ids", characterResponseList);
         return ResponseEntity.ok(responseDto);
     }
 
     @GetMapping("/search/name")
     public ResponseEntity<ResponseDto<CursorPage<CharacterResponse>>> findCharacterResponsesByCharacterName(@RequestParam("name") String name, @RequestParam(value = "cursorId", required = false) Long cursorId, @RequestParam int size) {
-        CursorPage<CharacterResponse> characterResponses = characterService.findAllByName(name, cursorId, size);
+        CursorPage<CharacterResponse> characterResponses = characterQueryService.findAllByName(name, cursorId, size);
         ResponseDto<CursorPage<CharacterResponse>> responseDto = HttpUtil.success("find characters", characterResponses);
         return ResponseEntity.ok(responseDto);
     }
