+++ src/main/java/windeath44/server/memorial/domain/character/service/usecase/MemorializingCharacterUseCase.java
@@ -2,7 +2,8 @@ package windeath44.server.memorial.domain.character.service.usecase;
 
 import windeath44.server.memorial.domain.character.mapper.CharacterMapper;
 import windeath44.server.memorial.domain.character.model.Character;
-import windeath44.server.memorial.domain.character.service.CharacterService;
+import windeath44.server.memorial.domain.character.service.CharacterCommandService;
+import windeath44.server.memorial.domain.character.service.CharacterQueryService;
 import windeath44.server.memorial.global.infrastructure.KafkaProducer;
 import com.example.avro.CharacterAvroSchema;
 import com.example.avro.MemorialAvroSchema;
@@ -15,18 +16,20 @@ import org.springframework.transaction.annotation.Transactional;
 @RequiredArgsConstructor
 @Slf4j
 public class MemorializingCharacterUseCase {
-  private final CharacterService characterService;
+  private final CharacterQueryService characterQueryService;
+  private final CharacterCommandService characterCommandService;
+
   private final KafkaProducer kafkaProducer;
   private final CharacterMapper characterMapper;
 
   @Transactional
   public void memorializing(MemorialAvroSchema memorialAvroSchema) {
-    Character character = characterService.findCharacterById(memorialAvroSchema.getCharacterId());
+    Character character = characterQueryService.findCharacterById(memorialAvroSchema.getCharacterId());
     CharacterAvroSchema characterAvroSchema = characterMapper.toCharacterAvroSchema(character, memorialAvroSchema);
 
     try {
       Long characterId = characterAvroSchema.getCharacterId();
-      characterService.memorializing(characterId);
+      characterCommandService.memorializing(characterId);
       kafkaProducer.send("character-memorialized-response", characterAvroSchema);
     } catch (Exception e) {
       log.error("memorializing character fail. characterId: {}", characterAvroSchema.getCharacterId());
